+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 migration_tools 7.x-2.x  ** - ** - ****
 -----------------------------------------------
+* Add QpHtml::removeComment()
+* Move some basic cleaning calls into SourceParser\HtmlBase.
 
 
 migration_tools 7.x-2.1  May 27, 2016